Implemented Spatial Domain Filters

+++ DigitalImageProcessing/ImageProcessingGUI/ImageOperations/AffineTransformation.py
@@ -8,7 +8,7 @@ def translate(image, delta_x=0, delta_y=0):
         [0, 1, delta_y]]
     )
     r, c = image.shape[:2]
-    return cv2.warpAffine(image, T, (c,r))
+    return np.uint8(cv2.warpAffine(image, T, (c,r)))
 
 def rotate(image, theta=0, clockwise=False, scale='auto'):
     theta = math.radians(theta%360)
@@ -23,7 +23,7 @@ def rotate(image, theta=0, clockwise=False, scale='auto'):
         [[a, b, (1-a)*center_x-b*center_y], 
         [-b, a, b*center_x+(1-a)*center_y]]
     )
-    return cv2.warpAffine(image, T, (c,r))
+    return np.uint8(cv2.warpAffine(image, T, (c,r)))
 
 def rescale(image, scale_x=1, scale_y=1):
     T = np.float32(
@@ -31,7 +31,7 @@ def rescale(image, scale_x=1, scale_y=1):
         [0, scale_y, 0]]
     )
     r, c = image.shape[:2]
-    return cv2.warpAffine(image, T, (c,r))
+    return np.uint8(cv2.warpAffine(image, T, (c,r)))
 
 def shear(image, shear_x=0, shear_y=0):
     T = np.float32(
@@ -39,7 +39,7 @@ def shear(image, shear_x=0, shear_y=0):
         [shear_x, 1, 0]]
     )
     r, c = image.shape[:2]
-    return cv2.warpAffine(image, T, (c,r))
+    return np.uint8(cv2.warpAffine(image, T, (c,r)))
 
 def resize(image, new_x, new_y, interpolation='NEAREST'):
     inter_map = {
@@ -47,4 +47,4 @@ def resize(image, new_x, new_y, interpolation='NEAREST'):
         'BILINEAR': cv2.INTER_LINEAR,
         'CUBIC': cv2.INTER_CUBIC
     }
-    return cv2.resize(image, (new_x, new_y), interpolation=inter_map[interpolation])
+    return np.uint8(cv2.resize(image, (new_x, new_y), interpolation=inter_map[interpolation]))
